Added dependence on max radius for moment of inertia calculation

+++ read/args.c
@@ -22,7 +22,7 @@ void flags(int argc, char *argv[])
 				break;
 			case 'i':		// Calculate moment of inertia tensor
 				iflag= !iflag;
-				radius=atof(optarg);	
+				radius=atoi(optarg);	
 				break;
 		}
     	}
